Microscopic virial stress with voronoi centre on cells.
Fixed bug in halfedge ordering which was affecting the contact length force.

Finally settled on a method to calculate virial stress of a cell. Also a slightly unresasonable method for calculating virial on vertices.

virial pressure goes to zero for gamma and lambda nonzero for hex test case.

Theory mistake in the stress calculation. Replaced an actual length with a proportion of the bond in the smoothing region, virial stress is updated.

simple microscopic expression for cell stress.

Fast and correct virial stress calculation

A few constructmesh routines.